Fix(JSON): Padding characters
Change padding characters from = tp + and vice versa when encrypted data are saved to the file / retrieved from it. This intends to avoid any confusion when the json is read and splitted on "=" characters. First split the data on "=", then convert the "+" padding characters back to "=", then decrypt the message.

Next step: check why all servers do not appear in "chose server" dialog.